Training/requirements.txt
numpy
pandas
scikit-learn
SQLAlchemy
xgboost
tensorflow
python-dotenv
